Update migration to prevent duplicate key error in Pimcore 10->11 upgrades
Changes in this pull request
Resolves pimcore#14995

Additional info
Pimcore 10 to 11 upgrades sometimes get stuck being unable to complete because the migrations are unable to run. The issue appears to be that this migration tries to add a row to a table which already exists. This was reported a couple times on the issue above which was closed due to inactivity and not because it was resolved. We have also experienced it with multiple of our clients doing Pimcore 11 upgrades where manually deleting the offending database row unsticks the migrations.